+++ TemiEsame/Gennaio22/Es2Float/main.c
@@ -1,3 +1,5 @@
+
+
 #include <stdio.h>
 #include <stdlib.h>
 #include "interfaccia2.h"
@@ -5,13 +7,13 @@
 
 int main()
 {
-    int array[] = {1, 2};
-    int *p = array;
-    vettore_pointer v1 = makeVettore(p, sizeof(array) / sizeof(int));
+    float array[] = {1, 2};
+    float *p = array;
+    vettore_pointer v1 = makeVettore(p, sizeof(array) / sizeof(float));
 
-    int array2[] = {1, 2};
-    int *p2 = array2;
-    vettore_pointer v2 = makeVettore(p2, sizeof(array2) / sizeof(int));
+    float array2[] = {1, 2};
+    float *p2 = array2;
+    vettore_pointer v2 = makeVettore(p2, sizeof(array2) / sizeof(float));
 
     printf("%s \n", toString(v1));
     printf("%s \n", toString(v2));
